Move tests to replica/leader mods
