--- node/actors/bft/src/tests/mod.rs
@@ -1,2 +1,0 @@
-mod network;
-mod unit;
